# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AgroTrack API",
    "description": "API para el sistema de gestión ganadera AgroTrack",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo AgroTrack",
      "email": "info@agrotrack.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Produccion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "animalId": {
            "type": "integer",
            "format": "int64"
          },
          "tipoProduccion": {
            "type": "string",
            "enum": ["leche", "carne"]
          },
          "cantidad": {
            "type": "number",
            "format": "float"
          },
          "fechaRegistro": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/produccion": {
      "get": {
        "tags": ["Producción"],
        "summary": "Obtener todos los registros de producción",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "animalId",
            "schema": {
              "type": "integer"
            },
            "description": "Filtrar por ID del animal"
          },
          {
            "in": "query",
            "name": "tipoProduccion",
            "schema": {
              "type": "string",
              "enum": ["leche", "carne"]
            },
            "description": "Filtrar por tipo de producción"
          },
          {
            "in": "query",
            "name": "fechaInicio",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Fecha de inicio para filtrar"
          },
          {
            "in": "query",
            "name": "fechaFin",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Fecha de fin para filtrar"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Ordenar resultados (ej. cantidad:desc)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de registros de producción",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Produccion"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "post": {
        "tags": ["Producción"],
        "summary": "Crear un nuevo registro de producción",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["animalId", "tipoProduccion", "cantidad"],
                "properties": {
                  "animalId": {
                    "type": "integer"
                  },
                  "tipoProduccion": {
                    "type": "string",
                    "enum": ["leche", "carne"]
                  },
                  "cantidad": {
                    "type": "number",
                    "format": "float"
                  },
                  "fechaRegistro": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registro de producción creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registro de producción creado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Produccion"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Animal no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
    "/api/produccion/{id}": {
      "get": {
        "tags": ["Producción"],
        "summary": "Obtener un registro de producción por ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del registro de producción"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro de producción",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Produccion"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Registro no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "put": {
        "tags": ["Producción"],
        "summary": "Actualizar un registro de producción",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del registro de producción"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipoProduccion": {
                    "type": "string",
                    "enum": ["leche", "carne"]
                  },
                  "cantidad": {
                    "type": "number",
                    "format": "float"
                  },
                  "fechaRegistro": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registro de producción actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registro de producción actualizado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Produccion"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Registro no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "delete": {
        "tags": ["Producción"],
        "summary": "Eliminar un registro de producción",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del registro de producción"
          }
        ],
        "responses": {
          "200": {
            "description": "Registro de producción eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registro de producción eliminado exitosamente"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Registro no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
    "/api/produccion/reporte": {
      "get": {
        "tags": ["Producción"],
        "summary": "Generar reporte de producción",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "periodo",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["diario", "semanal", "mensual"]
            },
            "description": "Periodo para el reporte"
          },
          {
            "in": "query",
            "name": "tipoProduccion",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["leche", "carne"]
            },
            "description": "Tipo de producción para el reporte"
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte de producción generado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "periodo": {
                          "type": "string",
                          "example": "diario"
                        },
                        "tipoProduccion": {
                          "type": "string",
                          "example": "leche"
                        },
                        "reporte": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "periodo": {
                                "type": "string",
                                "example": "2023-05-20"
                              },
                              "totalProduccion": {
                                "type": "number",
                                "example": 53.3
                              },
                              "cantidadAnimales": {
                                "type": "integer",
                                "example": 1
                              },
                              "promedioProduccion": {
                                "type": "number",
                                "example": 26.65
                              }
                            }
                          }
                        },
                        "alerta": {
                          "type": "string",
                          "nullable": true,
                          "example": "¡Alerta! No se ha registrado producción de leche en los últimos 7 días."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    }
  }
}